[update] split js file
